0.9.62

+++ src/Nalix.Common/Constants/PacketConstants.cs
@@ -10,15 +10,15 @@ public static class PacketConstants
     /// The threshold size (in bytes) for using heap-based memory allocation.
     /// This value represents the maximum size for which memory should be allocated from the heap instead of the stack.
     /// </summary>
-    public const int HeapAllocLimit = 1024;
+    public const System.Int32 HeapAllocLimit = 1024;
 
     /// <summary>
     /// Maximum stack allocation size (in bytes).
     /// </summary>
-    public const ushort StackAllocLimit = 512;
+    public const System.Int32 StackAllocLimit = 512;
 
     /// <summary>
     /// The maximum allowed packet size (in bytes), which is 64KB (65535 bytes).
     /// </summary>
-    public const ushort PacketSizeLimit = 0xFFFF;
+    public const System.Int32 PacketSizeLimit = 0xFFFF;
 }
